requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
matplotlib>=3.8.0
joblib>=1.4.0
gradio>=5.0.0
langdetect>=1.0.9
dvc>=3.0.0
python-dotenv>=1.0.1
openai>=1.30.0
datasets>=2.19.0
seaborn>=0.13.2
